Move BaseContext to frontend module
Also moves its dependencies
