+++ community/cypher/cypher-compiler-3.2/src/test/scala/org/neo4j/cypher/internal/compiler/v3_2/NotImplementedPlanContext.scala
@@ -20,6 +20,7 @@
 package org.neo4j.cypher.internal.compiler.v3_2
 
 import org.neo4j.cypher.internal.compiler.v3_2.spi._
+import org.neo4j.cypher.internal.frontend.v3_2.phases.InternalNotificationLogger
 import org.neo4j.kernel.api.constraints.UniquenessConstraint
 
 class NotImplementedPlanContext extends PlanContext {
