+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/ParsedQuery.scala
@@ -19,7 +19,7 @@
  */
 package org.neo4j.cypher.internal
 
-import org.neo4j.cypher.internal.compiler.v3_2.CompilationPhaseTracer
+import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer
 import org.neo4j.cypher.internal.spi.v3_2.TransactionalContextWrapper
 
 import scala.util.Try
