+++ community/cypher/frontend-3.2/src/main/java/org/neo4j/cypher/internal/frontend/v3_2/phases/CompilationPhaseTracer.java
@@ -17,22 +17,38 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compiler.v3_2.phases
+package org.neo4j.cypher.internal.frontend.v3_2.phases;
 
-import org.neo4j.cypher.internal.frontend.v3_2.phases.CompilationPhaseTracer.CompilationPhase.PARSING
-import org.neo4j.cypher.internal.frontend.v3_2.ast.Statement
-import org.neo4j.cypher.internal.frontend.v3_2.parser.CypherParser
-import org.neo4j.cypher.internal.frontend.v3_2.phases.BaseContext
+public interface CompilationPhaseTracer
+{
+    enum CompilationPhase
+    {
+        PARSING,
+        DEPRECATION_WARNINGS,
+        SEMANTIC_CHECK,
+        AST_REWRITE,
+        LOGICAL_PLANNING,
+        CODE_GENERATION,
+        PIPE_BUILDING,
+    }
 
-case object Parsing extends Phase[BaseContext] {
-  private val parser = new CypherParser
+    CompilationPhaseEvent beginPhase( CompilationPhase phase );
 
-  override def process(in: CompilationState, ignored: BaseContext): CompilationState =
-    in.copy(maybeStatement = Some(parser.parse(in.queryText, in.startPosition)))
+    interface CompilationPhaseEvent extends AutoCloseable
+    {
+        @Override
+        void close();
+    }
 
-  override val phase = PARSING
+    CompilationPhaseTracer NO_TRACING = new CompilationPhaseTracer()
+    {
+        @Override
+        public CompilationPhaseEvent beginPhase( CompilationPhase phase )
+        {
+            return NONE_PHASE;
+        }
+    };
 
-  override val description = "parse text into an AST object"
-
-  override def postConditions: Set[Condition] = Set(Contains[Statement])
+    CompilationPhaseEvent NONE_PHASE = () -> {
+    };
 }
